+++ src/extension.ts
@@ -1,104 +1,71 @@
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-
-import * as vscode from 'vscode';
-import { LanguageClient } from 'vscode-languageclient/node';
-import { registerLogger, traceError, traceLog, traceVerbose } from './common/log/logging';
 import {
-    checkVersion,
-    getInterpreterDetails,
-    initializePython,
-    onDidChangePythonInterpreter,
-    resolveInterpreter,
-} from './common/python';
-import { restartServer } from './common/server';
-import { checkIfConfigurationChanged, getInterpreterFromSetting } from './common/settings';
-import { loadServerDefaults } from './common/setup';
-import { getLSClientTraceLevel } from './common/utilities';
-import { createOutputChannel, onDidChangeConfiguration, registerCommand } from './common/vscodeapi';
+  LinterGetOffensesFunction,
+  LinterParseFixOutputFunction,
+  LinterOffense,
+  LinterOffenseSeverity,
+} from 'vscode-linter-api';
+import { Uri } from 'vscode';
 
-let lsClient: LanguageClient | undefined;
-export async function activate(context: vscode.ExtensionContext): Promise<void> {
-    // This is required to get server name and module. This should be
-    // the first thing that we do in this extension.
-    const serverInfo = loadServerDefaults();
-    const serverName = serverInfo.name;
-    const serverId = serverInfo.module;
-
-    // Setup logging
-    const outputChannel = createOutputChannel(serverName);
-    context.subscriptions.push(outputChannel, registerLogger(outputChannel));
-
-    const changeLogLevel = async (c: vscode.LogLevel, g: vscode.LogLevel) => {
-        const level = getLSClientTraceLevel(c, g);
-        await lsClient?.setTrace(level);
-    };
-
-    context.subscriptions.push(
-        outputChannel.onDidChangeLogLevel(async (e) => {
-            await changeLogLevel(e, vscode.env.logLevel);
-        }),
-        vscode.env.onDidChangeLogLevel(async (e) => {
-            await changeLogLevel(outputChannel.logLevel, e);
-        }),
-    );
-
-    // Log Server information
-    traceLog(`Name: ${serverInfo.name}`);
-    traceLog(`Module: ${serverInfo.module}`);
-    traceVerbose(`Full Server Info: ${JSON.stringify(serverInfo)}`);
-
-    const runServer = async () => {
-        const interpreter = getInterpreterFromSetting(serverId);
-        if (interpreter && interpreter.length > 0 && checkVersion(await resolveInterpreter(interpreter))) {
-            traceVerbose(`Using interpreter from ${serverInfo.module}.interpreter: ${interpreter.join(' ')}`);
-            lsClient = await restartServer(serverId, serverName, outputChannel, lsClient);
-            return;
-        }
-
-        const interpreterDetails = await getInterpreterDetails();
-        if (interpreterDetails.path) {
-            traceVerbose(`Using interpreter from Python extension: ${interpreterDetails.path.join(' ')}`);
-            lsClient = await restartServer(serverId, serverName, outputChannel, lsClient);
-            return;
-        }
-
-        traceError(
-            'Python interpreter missing:\r\n' +
-                '[Option 1] Select python interpreter using the ms-python.python.\r\n' +
-                `[Option 2] Set an interpreter using "${serverId}.interpreter" setting.\r\n` +
-                'Please use Python 3.7 or greater.',
-        );
-    };
-
-    context.subscriptions.push(
-        onDidChangePythonInterpreter(async () => {
-            await runServer();
-        }),
-        onDidChangeConfiguration(async (e: vscode.ConfigurationChangeEvent) => {
-            if (checkIfConfigurationChanged(e, serverId)) {
-                await runServer();
-            }
-        }),
-        registerCommand(`${serverId}.restart`, async () => {
-            await runServer();
-        }),
-    );
-
-    setImmediate(async () => {
-        const interpreter = getInterpreterFromSetting(serverId);
-        if (interpreter === undefined || interpreter.length === 0) {
-            traceLog(`Python extension loading`);
-            await initializePython(context.subscriptions);
-            traceLog(`Python extension loaded`);
-        } else {
-            await runServer();
-        }
-    });
+interface ProspectorOffense {
+  messages: ProspectorMessage[];
 }
 
-export async function deactivate(): Promise<void> {
-    if (lsClient) {
-        await lsClient.stop();
+interface ProspectorMessage {
+  source: string;
+  code: string;
+  location: ProspectorLocation;
+  message: string;
+  docUrl?: string;
+  fixable?: boolean;
+}
+
+interface ProspectorLocation {
+  path: string;
+  module: string;
+  function: string;
+  line: number;
+  character: number;
+  lineEnd?: number;
+  characterEnd?: number;
+}
+
+const messageRegex = /(?<message>.*) \[See: (?<url>.*)\]/;
+
+const offenseSeverity: { [key: string]: LinterOffenseSeverity } = {
+  convention: LinterOffenseSeverity.warning,
+  warning: LinterOffenseSeverity.error,
+};
+
+export const getOffenses: LinterGetOffensesFunction = ({ stdout, uri }: { stdout: string; uri: Uri }) => {
+  const result: ProspectorOffense = JSON.parse(stdout);
+  const offenses: LinterOffense[] = [];
+
+  result.messages.forEach((message: ProspectorMessage) => {
+    let messageString = message.message;
+    let docsUrl = '';
+    const match = message.message.match(messageRegex);
+    if (match) {
+      messageString = match.groups!.message;
+      docsUrl = match.groups!.url;
     }
-}
+
+    offenses.push({
+      uri,
+      lineStart: Math.max(0, message.location.line - 1),
+      columnStart: Math.max(0, message.location.character - 1),
+      lineEnd: Math.max(0, message.location.lineEnd || message.location.line - 1),
+      columnEnd: Math.max(0, message.location.characterEnd || message.location.character),
+      code: `${message.source}:${message.code}`,
+      message: messageString,
+      severity: LinterOffenseSeverity.error,
+      source: 'prospector',
+      correctable: message.fixable || false,
+      docsUrl: message.docUrl || docsUrl,
+    });
+  });
+
+  return Promise.resolve(offenses);
+};
+
+export const parseFixOutput: LinterParseFixOutputFunction = ({ stdout }: { stdout: string }) =>
+  Promise.resolve(stdout);